+++ packages/cli/configuration-loader/src/generators-yml/convertGeneratorsConfiguration.ts
@@ -731,6 +731,7 @@ async function convertGenerator({
         keywords: generator.keywords,
         smartCasing: generator["smart-casing"] ?? true,
         smartCasingDigitWordBoundary: generator["smart-casing-digit-word-boundary"] ?? false,
+        additionalAcronyms: generator["additional-acronyms"],
         disableExamples: generator["disable-examples"] ?? false,
         absolutePathToLocalOutput:
             generator.output?.location === "local-file-system"